breaking the appcomponent into smaller component for easier managing
