+++ WarehouseManagerS/client/src/app/ItemsList/item-list.component.css
@@ -1,0 +1,27 @@
+body {
+}
+.container {
+  margin-top: 5px;
+}
+
+table {
+  width: 100%;
+  border-collapse: collapse;
+}
+
+
+th, td {
+  padding: 8px 12px;
+  border: 5px solid #ddd;
+}
+
+th {
+  background-color: #f4f4f4;
+}
+
+h1 {
+  text-align: center;
+}
+
+button {
+}
