+++ WarehouseManagerS/client/src/app/ItemsList/item-list.component.html
@@ -1,0 +1,47 @@
+<div class="container">
+  <div class="toolbar">
+    <button type="button" (click)="addItem()">
+      <i class="fas fa-plus"></i> Add
+    </button>
+  </div>
+  <form [formGroup]="form">
+    <table>
+      <thead>
+        <tr>
+          <th>ID</th>
+          <th>Name</th>
+          <th>Quantity</th>
+          <th>Category</th>
+          <th align="center">Actions</th>
+        </tr>
+      </thead>
+      <tbody formArrayName="items">
+        <tr *ngFor="let item of itemsControls; let i = index" [formGroupName]="i">
+          <td>{{ i + 1 }}</td>
+          <td>
+            <input formControlName="name" />
+          </td>
+          <td>
+            <input formControlName="quantity" type="number" />
+          </td>
+          <td>
+            <select formControlName="category">
+              <option *ngFor="let category of categories" [value]="category">{{ category }}</option>
+            </select>
+          </td>
+          <td class="actions-cell">
+            <button type="button" *ngIf="!item.value.isEditing" (click)="startEdit(i)">
+              <i class="fas fa-edit"></i>
+            </button>
+            <button type="button" *ngIf="item.value.isEditing" (click)="saveItem(i)">
+              <i class="fas fa-save"></i>
+            </button>
+            <button type="button" (click)="deleteItem(i)">
+              <i class="fas fa-trash"></i>
+            </button>
+          </td>
+        </tr>
+      </tbody>
+    </table>
+  </form>
+</div>
