--- WarehouseManagerS/client/src/app/ItemsList/itemList.component.css
@@ -1,2 +1,0 @@
-﻿body {
-}
