--- WarehouseManagerS/client/src/app/ItemsList/itemList.component.html
@@ -1,10 +1,0 @@
-﻿<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8" />
-    <title></title>
-</head>
-<body>
-
-</body>
-</html>
