--- WarehouseManagerS/client/src/app/ItemsList/itemList.component.ts
@@ -1,9 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-list',
-  templateUrl: './itemList.component.html'
-})
-export class itemList {
-
-}
