+++ WarehouseManagerS/client/src/app/app-routing.module.ts
@@ -1,7 +1,17 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { HomeComponent } from './home/home.component';
+import { ItemListComponent } from './ItemsList/item-list.component';
+import {NavComponent } from './nav/nav.component'
+// Importuj inne komponenty według potrzeb
 
-export const routes: Routes = [];
+
+const routes: Routes = [
+  { path: '', component: HomeComponent }, // Ścieżka domyślna (np. strona główna)
+  { path: 'items', component: ItemListComponent }, // Ścieżka do listy przedmiotów
+  // Dodaj inne ścieżki według potrzeb
+];
+
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
