+++ WarehouseManagerS/client/src/app/app.module.ts
@@ -1,18 +1,18 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
-import { ReactiveFormsModule } from '@angular/forms';
-import { FormsModule } from '@angular/forms';
+import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { AppComponent } from './app.component';
 import { ItemService } from './_services/item.service';
 import { NavComponent } from './nav/nav.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { BsDropdownModule } from 'ngx-bootstrap/dropdown';
+import { ItemListComponent } from './ItemsList/item-list.component';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    NavComponent
+  declarations: [     // Dodaj AppComponent tutaj
+    NavComponent,
+    ItemListComponent
   ],
   imports: [
     BrowserModule,
@@ -20,9 +20,13 @@ import { BsDropdownModule } from 'ngx-bootstrap/dropdown';
     ReactiveFormsModule,
     FormsModule,
     BrowserAnimationsModule,
+    CommonModule,
+    FormsModule,
+    ReactiveFormsModule,
+    RouterModule,
     BsDropdownModule.forRoot()
   ],
   providers: [ItemService],
-  bootstrap: [AppComponent]
+  bootstrap: []
 })
 export class AppModule { }
