+++ WarehouseManagerS/client/src/app/home/home.component.ts
@@ -3,7 +3,6 @@ import { RegisterComponent } from '../register/register.component';
 
 @Component({
   selector: 'app-home',
-  standalone: true,
   imports: [RegisterComponent],
   templateUrl: './home.component.html',
   styleUrl: './home.component.css'
